Added template for a WfMS plugin
Added a template for a WfMS plugin, so that we have a starting project
for the time when we have to create new ones.

The plugin is a maven project and is already set up for packaging the
plugin with the intended structure and manifest file in a .zip archive.